refactor: update import paths for ControllerGenerator and add generator class
